Tested MechMaid progress pkg as a library module.
Everything seems functional. Tested progressable content frame in two
places. Also tested the default progress view by adding a custom
layout within the app, which was automatically detected and used. Would
be awesome if we could allow the user to specify the name of the default
instead of having to name it according to our convention.